refactor(ari): extract `Event` to separate module
