+++ arirs/src/lib.rs
@@ -1,10 +1,3 @@
-use channel::{
-    ChannelCreated, ChannelDestroyed, ChannelDialplan, ChannelDtmfReceived, ChannelHangupRequest, ChannelStateChange, ChannelVarset,
-    StasisEnd, StasisStart,
-};
-use device::DeviceStateChanged;
-use serde::{Deserialize, Serialize};
-
 pub mod bridge;
 pub mod channel;
 pub mod client;
@@ -17,19 +10,5 @@ pub mod variable;
 mod error;
 pub use error::{AriError, Result};
 
-#[derive(Serialize, Deserialize, Debug)]
-#[serde(tag = "type")]
-pub enum Event {
-    StasisStart(StasisStart),
-    StasisEnd(StasisEnd),
-    ChannelCreated(ChannelCreated),
-    ChannelDestroyed(ChannelDestroyed),
-    ChannelVarset(ChannelVarset),
-    ChannelHangupRequest(ChannelHangupRequest),
-    ChannelDialplan(ChannelDialplan),
-    ChannelStateChange(ChannelStateChange),
-    ChannelDtmfReceived(ChannelDtmfReceived),
-    DeviceStateChanged(DeviceStateChanged),
-    #[serde(other)]
-    Unknown,
-}
+mod event;
+pub use event::Event;
